% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mint_utils.R
\name{monthly_category_sum}
\alias{monthly_category_sum}
\title{get monthly sum of spending by category}
\usage{
monthly_category_sum(
  transactions,
  config_file,
  start_date,
  include_outlier = FALSE
)
}
\arguments{
\item{transactions}{df of mint transactions}

\item{start_date}{date to keep transactions}

\item{include_outlier}{whether to remove outliers as defined in user config}
}
\value{

}
\description{
get monthly sum of spending by category
}
\examples{
monthly_category_sum(transactions, config_file, start_date, include_outlier = FALSE)
}
